Fix bugs in the handling of YAML and TAG directives
Both directives were broken by a typo in the Scanner.  Hilariously, each
was then broken individually in different ways.

The YAML directive handling was broken due to the crazy formatting of
the exception throwing code.  The fix is stylistically poor but is
consistent with the rest of the file.

The TAG directive handling was broken due to a typo when extracting the
values from the token.

Fixes #24